Improve handling of configuration data
Update the uploading methods to be more tolerant of the data passed to
them.  They will accept paths to files, directories, or arrays of files.
Anything passed in that is unable to be located, such as an invalid path, will
be skipped instead of failing the run.

Additionally cleaned up the Vagrant file to point to included fixtures.

Attempted to get the UI methods to display [Chef Zero] instead of [default] to
no avail.

The gemspec requires JSON 1.7.7 due to issues with Bundler, this should only
affect development workflows.