Fix issues with Chef-Zero and Vagrant reload action
This commit rolls back most of the previous work around switching to
using the Chef Zero gem directly.  This issue with using it to launch
the server is that it does not do a double fork to act as a standalone
service.  What this implies is that once the vagrant command finishes
the Chef Zero server, which is a child process of the vagrant command,
will be shut down as well.

It seems that binding after the VBox interface take down causes issues
now, so I have switched it to perform just before it instead (yay
Vagrant vagaries).