Find provisioners in backwards compatible manner
Default to checking the provisioner type attribute if it exists,
otherwise fallback to the old style name attribute.

Vagrant changed provisioners definition for >=1.7 to rename the name
attribute to type, and now uses the name attribute for the unique id
which will never match across multiple provisioners of the same type.

By checking to see if the attribute is defined before comparing we can
support all vagrant versions from 1.4+.