Use bump gem and generate unique versions from git
Use the bump gem to manage updating the version, lock file and tags
together, ensure the runtime version is retrieved from the built spec
so that we can generate pre-release versions directly from git.

Use the Rakefile to modify the build task installed by gem_tasks so that
the git version is only applied to when using the build and install
tasks. This ensures that when using the bundle to install the
dependencies in deployment mode for testing that bundle can find the
correct version of the project.